requirements.txt
openai
streamlit
python-dotenv
langchain
langchain_community
langchain_openai
pypdf